# backend/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Fluent Interview API",
    "description": "API untuk sistem wawancara berbasis AI dan manajemen pengguna FLUENT.",
    "version": "1.0.1",
    "contact": {
      "name": "Fluent Support",
      "email": "contact@fluentapp.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Server"
    },
    {
      "url": "https://api.fluentapp.com",
      "description": "Production Server (Contoh)"
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Registrasi Pengguna Baru (via API)",
        "description": "Mendaftarkan pengguna baru dengan email, username, password, dan OTP jika alur Anda menggunakannya. Endpoint ini mungkin untuk API langsung atau setelah verifikasi OTP.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegistrationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registrasi berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRegistrationSuccessResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login Pengguna (API)",
        "description": "Mengautentikasi pengguna dengan email dan password, mengembalikan token JWT.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserLoginSuccessResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Perbarui Access Token (API)",
        "description": "Menggunakan refresh token untuk mendapatkan access token baru.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token berhasil diperbarui",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenSuccessResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserBase": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "username": { "type": "string", "example": "john_doe" }
        }
      },
      "UserRegistrationRequest": {
        "allOf": [
          { "$ref": "#/components/schemas/UserBase" },
          {
            "type": "object",
            "properties": {
              "password": { "type": "string", "format": "password", "example": "strongpassword123" },
              "gender": { "type": "string", "enum": ["Male", "Female", "Not specified"], "example": "Male", "nullable": true },
              "occupation": { "type": "string", "example": "Software Engineer", "nullable": true },
              "otp": { "type": "string", "description": "OTP jika diperlukan oleh alur registrasi Anda", "example":"123456", "nullable": true}
            },
            "required": ["password"]
          }
        ]
      },
      "UserLoginRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email", "example": "user@example.com" },
          "password": { "type": "string", "format": "password", "example": "strongpassword123" }
        },
        "required": ["email", "password"]
      },
      "UserData": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "objectid", "example": "60d0fe4f5311236168a109ca" },
          "email": { "type": "string", "format": "email" },
          "username": { "type": "string" },
          "gender": { "type": "string", "nullable": true },
          "occupation": { "type": "string", "nullable": true },
          "is_active": { "type": "boolean" },
          "profile_picture": { "type": "string", "format": "url", "nullable": true },
          "auth_provider": { "type": "string", "enum": ["local", "google"], "example": "local" }
        }
      },
      "UserRegistrationSuccessResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "success" },
          "message": { "type": "string", "example": "User registered successfully" },
          "user": {
            "type": "object",
            "properties": {
                "id": { "type": "string", "format": "objectid" },
                "username": { "type": "string" },
                "email": { "type": "string", "format": "email" }
            }
          }
        }
      },
      "UserLoginSuccessResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "success" },
          "message": { "type": "string", "example": "Login successful" },
          "access_token": { "type": "string", "format": "jwt" },
          "refresh_token": { "type": "string", "format": "jwt" },
          "user": { "$ref": "#/components/schemas/UserData" }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "properties": {
          "refresh_token": { "type": "string", "format": "jwt" }
        },
        "required": ["refresh_token"]
      },
      "RefreshTokenSuccessResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "success" },
          "access_token": { "type": "string", "format": "jwt" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "fail" },
          "message": { "type": "string" }
        }
      },
      "ErrorResponseWithData": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "error" },
          "message": { "type": "string" },
          "data": { "type": "object", "additionalProperties": true }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Permintaan tidak valid atau data yang hilang.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "Unauthorized": {
        "description": "Autentikasi gagal atau token tidak valid/kedaluwarsa.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "Forbidden": {
        "description": "Akses ditolak, pengguna tidak memiliki izin.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "NotFound": {
        "description": "Sumber daya yang diminta tidak ditemukan.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "Conflict": {
        "description": "Konflik data, misalnya username/email sudah ada.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "TooManyRequests": {
        "description": "Terlalu banyak percobaan (misalnya, login).",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
      "InternalServerError": {
        "description": "Terjadi kesalahan internal pada server.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponseWithData" } } }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Masukkan token JWT dengan prefix 'Bearer '. Contoh: 'Bearer abcdef12345'"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key untuk mengakses endpoint tertentu."
      }
    }
  }
}